requirements.txt
python-okx>=0.3.0
sqlalchemy~=2.0.25
pyyaml~=6.0.1